+++ ProjectFlightSchool/ProjectFlightSchool/Client.h
@@ -113,7 +113,7 @@ void Client::HandlePkg( Package<T>* p )
 		case Net_Event::EV_UPDATE_HP:
 		{
 			EvPlayerID player = (EvPlayerID&)p->body.content;
-			IEventPtr E1( new Event_Remote_Player_Update_HP( player.ID, player.HP  ) );
+			IEventPtr E1( new Event_Remote_Player_Update_HP( player.ID, (float)player.HP  ) );
 			EventManager::GetInstance()->QueueEvent( E1 );
 		}
 			break;
